[fix] content fuel loading
