+++ layouts/Header/index.tsx
@@ -17,13 +17,17 @@ import { useCreatorAccount } from '@/hooks/useCreatorAccount';
 
 import LoadingConnectWallet from '@/components/Animation/LoadingConnectWallet';
 import ModalInviteCode from '@/components/InviteCode/ModalInviteCode';
+import { cairo, Contract, Provider } from 'starknet';
+import { ABIS } from '@/abis';
+import { CONTRACT_ADDRESS } from '@/utils/constants';
+import systemConfig from '@/config/systemConfig';
 
 const Header = () => {
   const { userAddress, prevConnector, isLoading, verifySignature } = useAuth();
   const { connectors, connect } = useConnect();
   const { isOpen, onClose, onOpen } = useDisclosure();
 
-  const { handleClearEventLog } = useCreatorAccount();
+  const { handleClearEventLog, handleSetPoint } = useCreatorAccount();
   const {
     address: addressWallet,
     status: statusWallet,
@@ -31,7 +35,17 @@ const Header = () => {
   } = useAccount();
 
   const dispatch = useDispatch();
+  const getUserPoint = async (userAddress: string) => {
+    const contractBlizt = new Contract(
+      ABIS.bliztABI,
+      CONTRACT_ADDRESS.BLIZT_POINT,
+      new Provider({ nodeUrl: systemConfig().RPC })
+    );
 
+    const data = await contractBlizt.getUserPoint(userAddress);
+    const fomatPoint = Number(cairo.uint256(data).low.toString());
+    handleSetPoint(fomatPoint);
+  };
   useEffect(() => {
     const handleChangeWallet = async () => {
       if (
@@ -43,6 +57,7 @@ const Header = () => {
         dispatch(setUserLoading(true));
         handleClearEventLog();
         await verifySignature(account);
+        await getUserPoint(addressWallet);
         dispatch(setUserLoading(false));
       } else if (
         addressWallet &&
@@ -53,6 +68,7 @@ const Header = () => {
         dispatch(setUserLoading(true));
         handleClearEventLog();
         await verifySignature(account);
+        await getUserPoint(addressWallet);
         dispatch(setUserLoading(false));
       }
     };
